pyproject.toml
[project]
name = "vaa-ai-travel-assistant"
version = "0.1.0"
description = "An AI Travel Assistant for Virgin Atlantic that offers recommendations for travel destinations, hotels, flights, and activities."
authors = [
    {name = "Chin Perera",email = "chinthanaperera@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "openai (>=1.97.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-ai (>=0.4.7,<0.5.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "langchain-core (>=0.3.72,<0.4.0)",
    "langchain-chroma (>=0.2.5,<0.3.0)",
    "streamlit (>=1.47.1,<2.0.0)",
    "logfire (>=4.0.0,<5.0.0)",
]

[tool.poetry]
packages = [
    { include = "app" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
pytest-mock = "^3.14.1"
httpx = "^0.28.1"
pytest-cov = "^6.2.1"
pylint = "^3.3.7"


